Explicit casts to (uint8_t)
kvldskey_mlen() will always return a value which fits into (uint8_t).  It
begins by setting
    size_t mlen = 0
and then increments it until
    mlen == kvldskey->(uint8_t len)
(or until the buffers do not match).  The definition of `struct kvldskey` is in
    lib/datastruct/kvldskey.h

+++ kvlds/btree_node.c
@@ -155,7 +155,7 @@ btree_node_mknode(struct btree * T, int type, int height, size_t nkeys,
 	if (type == NODE_TYPE_LEAF) {
 		N->u.pairs = pairs;
 		if ((N->nkeys > 0) && (pairs != NULL)) {
-			N->mlen_n = kvldskey_mlen(N->u.pairs[0].k,
+			N->mlen_n = (uint8_t)kvldskey_mlen(N->u.pairs[0].k,
 			    N->u.pairs[N->nkeys - 1].k);
 		} else {
 			N->mlen_n = 255;
